+++ Menu/NumerosPorParticipante.c
@@ -1,0 +1,10 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <locale.h>
+#include <string.h>
+#include <math.h>
+#include <time.h>
+
+void NumerosPorParticipante()
+{
+}
